+++ mdformat/renderer/_util.py
@@ -1,10 +1,11 @@
 import html.entities
 import re
-from typing import Callable, Sequence
-
-from markdown_it.token import Token
+from typing import TYPE_CHECKING
 
 from mdformat.renderer import _codepoints
+
+if TYPE_CHECKING:
+    from mdformat.renderer import TreeNode
 
 # Regex that finds character references.
 # The reference can be either
@@ -22,103 +23,25 @@ RE_CHAR_REFERENCE = re.compile(
 CONSECUTIVE_KEY = "number"
 
 
-class MARKERS:
-    """A container for markers for unprocessed Markdown.
-
-    Temporary markers that can be used for markdown that is not yet
-    fully processed.
-    """
-
-    # Type ignore, because mypy thinks this is a method. This is not
-    # a method, just a function in the namespace.
-    def _make_autovalue_func() -> Callable[[], str]:  # type: ignore
-        """A closure to make a function that creates unique, even length MARKER
-        values.
-
-        "\x00" is invalid Markdown so a string that contains it can not
-        naturally exist in Markdown. This is why we prefix all markers
-        with it.
-        """
-        count = 0
-        max_val = 99
-        number_len = len(str(max_val))
-
-        def next_value() -> str:
-            nonlocal count
-            if count > max_val:
-                raise Exception("Tried to create too many marker values")
-            marker = "\x00" + str(count).rjust(number_len, "0")
-            count += 1
-            return marker
-
-        return next_value
-
-    _autovalue = _make_autovalue_func()
-
-    LIST_ITEM = _autovalue()
-    LIST_INDENT_FIRST_LINE = _autovalue()
-    LIST_INDENT = _autovalue()
-    # We add BLOCK_SEPARATOR instead of newlines at the end of every block.
-    # We convert it to newlines when closing
-    #   - a list item
-    #   - a list
-    #   - a blockquote
-    #   - the root document
-    BLOCK_SEPARATOR = _autovalue()
-
-
-def index_opening_token(tokens: Sequence[Token], closing_token_idx: int) -> int:
-    """Return index of an opening token.
-
-    Takes token stream and closing token index of a container block as
-    params. Returns index of the corresponding opening token.
-    """
-    closing_tkn = tokens[closing_token_idx]
-    assert closing_tkn.nesting == -1, "Cant find opening token for non closing token"
-    for i in reversed(range(closing_token_idx)):
-        opening_tkn_candidate = tokens[i]
-        if opening_tkn_candidate.level == closing_tkn.level:
-            return i
-    raise ValueError("Invalid token list. Opening token not found.")
-
-
-def find_opening_token(tokens: Sequence[Token], closing_token_idx: int) -> Token:
-    """Return an opening token.
-
-    Takes token stream and closing token index of a container block as
-    params. Returns the corresponding opening token.
-    """
-    opening_token_idx = index_opening_token(tokens, closing_token_idx)
-    return tokens[opening_token_idx]
-
-
-def is_tight_list(tokens: Sequence[Token], closing_token_idx: int) -> bool:
-    closing_tkn = tokens[closing_token_idx]
-    assert closing_tkn.type in {"bullet_list_close", "ordered_list_close"}
+def is_tight_list(node: "TreeNode") -> bool:
+    assert node.type_ in {"bullet_list", "ordered_list"}
 
     # The list has list items at level +1 so paragraphs in those list
-    # items must be at level +2
-    paragraph_level = closing_tkn.level + 2
-
-    for i in range(
-        index_opening_token(tokens, closing_token_idx) + 1, closing_token_idx
-    ):
-        if tokens[i].level != paragraph_level or tokens[i].type != "paragraph_open":
-            continue
-        is_tight = tokens[i].hidden
-        if not is_tight:
-            return False
+    # items must be at level +2 (grand children)
+    for child in node.children:
+        for grand_child in child.children:
+            if grand_child.type_ != "paragraph":
+                continue
+            is_tight = grand_child.opening.hidden
+            if not is_tight:
+                return False
     return True
 
 
-def is_tight_list_item(tokens: Sequence[Token], closing_token_idx: int) -> bool:
-    list_item_closing_tkn = tokens[closing_token_idx]
-    assert list_item_closing_tkn.type == "list_item_close"
-
-    for i in range(closing_token_idx, len(tokens)):
-        if tokens[i].level < list_item_closing_tkn.level:
-            return is_tight_list(tokens, i)
-    raise ValueError("List closing token not found")
+def is_tight_list_item(node: "TreeNode") -> bool:
+    assert node.type_ == "list_item"
+    assert node.parent is not None
+    return is_tight_list(node.parent)
 
 
 def longest_consecutive_sequence(seq: str, char: str) -> int:
@@ -137,12 +60,13 @@ def longest_consecutive_sequence(seq: str, char: str) -> int:
     return longest
 
 
-def is_text_inside_autolink(tokens: Sequence[Token], idx: int) -> bool:
-    assert tokens[idx].type == "text"
-    if idx == 0:
-        return False
-    previous_token = tokens[idx - 1]
-    return previous_token.type == "link_open" and previous_token.markup == "autolink"
+def is_text_inside_autolink(node: "TreeNode") -> bool:
+    assert node.type_ == "text"
+    return (
+        node.parent
+        and node.parent.type_ == "link"
+        and node.parent.opening.markup == "autolink"
+    )
 
 
 def removesuffix(string: str, suffix: str) -> str:
@@ -160,8 +84,8 @@ def maybe_add_link_brackets(link: str) -> str:
     return link
 
 
-def get_list_marker_type(tokens: Sequence[Token], closing_token_idx: int) -> str:
-    if tokens[closing_token_idx].type == "bullet_list_close":
+def get_list_marker_type(node: "TreeNode") -> str:
+    if node.type_ == "bullet_list":
         mode = "bullet"
         primary_marker = "-"
         secondary_marker = "*"
@@ -170,18 +94,85 @@ def get_list_marker_type(tokens: Sequence[Token], closing_token_idx: int) -> str
         primary_marker = "."
         secondary_marker = ")"
     consecutive_lists_count = 1
-    current_closing_idx = closing_token_idx
+    current = node
     while True:
-        assert current_closing_idx, "Closing token index can not be 0"
-        opening_token_idx = index_opening_token(tokens, current_closing_idx)
-        if opening_token_idx == 0:
+        previous_sibling = current.previous_sibling()
+        if previous_sibling is None:
             return primary_marker if consecutive_lists_count % 2 else secondary_marker
-        prev_idx = opening_token_idx - 1
-        prev_type = tokens[prev_idx].type
-        if (mode == "bullet" and prev_type == "bullet_list_close") or (
-            mode == "ordered" and prev_type == "ordered_list_close"
+        prev_type = previous_sibling.type_
+        if (mode == "bullet" and prev_type == "bullet_list") or (
+            mode == "ordered" and prev_type == "ordered_list"
         ):
             consecutive_lists_count += 1
-            current_closing_idx = prev_idx
+            current = previous_sibling
         else:
             return primary_marker if consecutive_lists_count % 2 else secondary_marker
+
+
+def escape_asterisk_emphasis(text: str) -> str:
+    """Escape asterisks to prevent unexpected emphasis/strong emphasis.
+
+    Currently we escape all asterisks unless both previous and next
+    character are Unicode whitespace.
+    """
+    # Fast exit to improve performance
+    if "*" not in text:
+        return text
+
+    escaped_text = ""
+
+    text_length = len(text)
+    for i, current_char in enumerate(text):
+        if current_char != "*":
+            escaped_text += current_char
+            continue
+        prev_char = text[i - 1] if (i - 1) >= 0 else None
+        next_char = text[i + 1] if (i + 1) < text_length else None
+        if (
+            prev_char in _codepoints.UNICODE_WHITESPACE
+            and next_char in _codepoints.UNICODE_WHITESPACE
+        ):
+            escaped_text += current_char
+            continue
+        escaped_text += "\\" + current_char
+
+    return escaped_text
+
+
+def escape_underscore_emphasis(text: str) -> str:
+    """Escape underscores to prevent unexpected emphasis/strong emphasis.
+    Currently we escape all underscores unless:
+
+    - Neither of the surrounding characters are one of Unicode whitespace,
+      start or end of line, or Unicode punctuation
+    - Both surrounding characters are Unicode whitespace
+    """
+    # Fast exit to improve performance
+    if "_" not in text:
+        return text
+
+    bad_neighbor_chars = (
+        _codepoints.UNICODE_WHITESPACE
+        | _codepoints.UNICODE_PUNCTUATION
+        | frozenset({None})
+    )
+    escaped_text = ""
+
+    text_length = len(text)
+    for i, current_char in enumerate(text):
+        if current_char != "_":
+            escaped_text += current_char
+            continue
+        prev_char = text[i - 1] if (i - 1) >= 0 else None
+        next_char = text[i + 1] if (i + 1) < text_length else None
+        if (
+            prev_char in _codepoints.UNICODE_WHITESPACE
+            and next_char in _codepoints.UNICODE_WHITESPACE
+        ) or (
+            prev_char not in bad_neighbor_chars and next_char not in bad_neighbor_chars
+        ):
+            escaped_text += current_char
+            continue
+        escaped_text += "\\" + current_char
+
+    return escaped_text
